+++ src/modules/free-block/index.tsx
@@ -2,14 +2,13 @@ import registerModule from '~/core/module';
 import { UnloaderContext } from '~/core/module-loader';
 import {
   addAsyncDayLoadedListener,
-  addDayLoadedListener,
   getAnnouncementWrap,
   getDayViewDateString,
   isEmptySchedule,
   to24Hr,
 } from '~/shared/schedule';
 import { isDaylightSavings, timeStringToDate } from '~/utils/date';
-import { createElement, insertCss, waitForLoad } from '~/utils/dom';
+import { createElement, insertCss } from '~/utils/dom';
 import { fetchData } from '~/utils/fetch';
 import log from '~/utils/log';