+++ src/modules/highlight-current-class/index.ts
@@ -18,13 +18,6 @@ function removeHighlight() {
   }
 }
 
-// const domQuery = () => (
-//   document.getElementById('accordionSchedules')
-//   && document.getElementById('accordionSchedules').children[0]
-//   && document.getElementById('accordionSchedules').children[0].children
-//   && document.getElementById('accordionSchedules').children[0].children.length
-// );
-
 const domQuery = () => document.getElementById('accordionSchedules')?.children[0]?.children?.length;
 
 async function highlight(blocks: HTMLCollection) {